% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bankLabels}
\alias{bankLabels}
\title{Dummy labels for bankData}
\format{A data frame with 16 rows and 3 variables:
\describe{
  \item{V1}{name of each variable in bankData}
  \item{V2}{label of each variable in bankData}
  \item{V3}{a dummy variable, corresponding to the row number}
  ...
}}
\usage{
bankLabels
}
\description{
A dataset containing the attribute labels also found in \code{\link{bankData}}.
This data set is used to illustrate the PrepLabels function and other label
functionality in the otvPlots package in R
}
\keyword{datasets}

