% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_old.R
\docType{package}
\name{otvPlots}
\alias{otvPlots}
\alias{otvPlots-package}
\title{Over time variable plots for predictive modeling (otvPlots)}
\description{
The otvPlots package uses data.table and ggplot2 packages to efficiently plot
time series aggregated from large datasets. Plots are optionally returned
ordered by correlation with date -- a natural starting point for anomaly
detection.  Plots are automatically labeled if a variable dictionary is
provided. Discrete and numeric variables are handled automatically. The
package can be used either interactively for adhoc creation of plots for
specifical variables, or via the included wrapper function for hands-off
monitoring reports.
}
\section{License}{

Copyright 2016 Capital One Services, LLC Licensed under the Apache License,
Version 2.0 (the "License"); you may not use this file except in compliance
with the License. You may obtain a copy of the  License at
http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law
or agreed to in writing, software distributed under the License is 
distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY 
KIND, either express or implied. See the License for the specific language 
governing permissions and limitations under the License.
}

\seealso{
\code{\link{PrepData}}

\code{\link{PrepLabels}}
}
