% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otvPlotsFunctions.R
\name{PlotCatRate}
\alias{PlotCatRate}
\title{Plot Category Frequency Rates Over Time}
\usage{
PlotCatRate(dataFl, myVar, dateGp, kCategories, newLevels, weightNm = NULL)
}
\arguments{
\item{dataFl}{Either the name of an object that can be converted using
as.data.table (e.g. a ' data frame) or a character string containing the name
of dataset that can be loaded using fread (e.g. a csv file). ' If dataset is
not in your working directory then \code{dataFl} must include (relative or 
absolute) path to file}

\item{myVar}{Name of the variable to be plotted}

\item{dateGp}{Name of the variable the time series plots should be grouped
by. Options are NULL, "weeks", "months", "quarters", "years". See
data.table::IDate. If NULL \code{dateNm} will be used.}

\item{kCategories}{Number of categories to plot in over time
rate chart for categorical data -- the K most numerous categories in the 
dataset will be plotted. Default is 3.}

\item{newLevels}{A vector of the 9 most frequent categories of \code{myVar},
in order of global frequency as calculated by PlotDiscreteVar}

\item{weightNm}{Name of variable containing row weights or NULL for no
weights (all rows recieve weight 1)}
}
\value{
A ggproto object with plotting the proportion of \code{myVar} per category over time, 
for the most frequent (globally) \code{kCategories} of \code{myVar}
}
\description{
Plot Category Frequency Rates Over Time
}
\section{License}{

Copyright 2016 Capital One Services, LLC Licensed under the Apache License,
Version 2.0 (the "License"); you may not use this file except in compliance
with the License. You may obtain a copy of the  License at
http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law
or agreed to in writing, software distributed under the License is 
distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY 
KIND, either express or implied. See the License for the specific language 
governing permissions and limitations under the License.
}
\examples{
data(bankData)
setDT(bankData)
PrepData(bankData, dateNm = "date", dateGp = "months", dateGpBp = "quarters", 
         weightNm = NULL)
newLevels = names(rev(sort(bankData[, table(job)])))
PlotCatRate(dataFl = bankData, myVar = "job", dateGp = "months", 
				kCategories = 3, newLevels = newLevels)
}

