% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otvPlotsFunctions.R
\name{PlotContVar}
\alias{PlotContVar}
\title{Boxplots and overtime plots of numeric variables}
\usage{
PlotContVar(myVar, dataFl, weightNm, dateGp, dateGpBp, skewOpt = NULL,
  kSample = NULL)
}
\arguments{
\item{myVar}{Name of the variable to be plotted}

\item{dataFl}{A data.table containing at least the following column names:
\code{myVar}, \code{weightNm}, \code{dateGp}, \code{dateGpBp}. Output of
PrepData}

\item{weightNm}{Name of variable containing row weights or NULL for no
weights (all rows recieve weight 1)}

\item{dateGp}{Name of the variable the time series plots should be grouped
by. Options are NULL, "weeks", "months", "quarters", "years". See
data.table::IDate. If NULL \code{dateNm} will be used.}

\item{dateGpBp}{Name of variable the boxplots should be grouped by. Same
options as \code{dateGp}. If NULL \code{dateGp} will be used.}

\item{skewOpt}{Either numeric constant or NULL. If numeric, say 5, the box
plots of variables whose skew exceeds 5 will be on a log10 scale if
possible. If negative, 3 will be used as the cutoff. Default is NULL (no 
transformation)}

\item{kSample}{Either NULL or integer. If a number, indicates the sample size
for both drawing boxplots and ordering numerical graphs by R2. For large
datasets, setting this to a reasonable (say 50K) value dramatically
increases processing speed. In larger datasets (e.g. > 10 percent system
memory), this parameter should not be set to NULL, or boxplots may take a
very long time to render. This setting has no impact on the accuracy of
time series plots.}
}
\value{
A grid of ggplot objects including a boxplot grouped by 
\code{dateGpBp}, a time series of  p1, p50 and p99 grouped by \code{dateGp},
a time series of mean and +-1 SD control limits grouped by \code{dateGp},
and a time series of missing and zerorates grouped by \code{dateGp}
}
\description{
Boxplots and overtime plots of numeric variables
}
\section{License}{

Copyright 2016 Capital One Services, LLC Licensed under the Apache License,
Version 2.0 (the "License"); you may not use this file except in compliance
with the License. You may obtain a copy of the  License at
http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law
or agreed to in writing, software distributed under the License is 
distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY 
KIND, either express or implied. See the License for the specific language 
governing permissions and limitations under the License.
}
\examples{
data(bankData)
setDT(bankData)
PrepData(bankData, dateNm = "date", dateGp = "months", dateGpBp = "years")
plot(PlotContVar("balance", bankData, NULL, "months", "years", 
                 skewOpt = NULL, kSample = NULL))

}

