% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otvPlotsFunctions.R
\name{PlotDiscreteVar}
\alias{PlotDiscreteVar}
\title{Frequency and proportion plots for discrete variables}
\usage{
PlotDiscreteVar(myVar, dataFl, weightNm = NULL, dateNm, dateGp,
  kCategories = 9, normBy = "time")
}
\arguments{
\item{myVar}{Name of the variable to be plotted}

\item{dataFl}{A data.table containing at least the following column names:
\code{myVar}, \code{weightNm}, \code{dateGp}, \code{dateGpBp}. Output of
PrepData}

\item{weightNm}{Name of variable containing row weights or NULL for no
weights (all rows recieve weight 1)}

\item{dateNm}{Name of column containing \code{date} variable}

\item{dateGp}{Name of the variable the time series plots should be grouped
by. Options are NULL, "weeks", "months", "quarters", "years". See
data.table::IDate. If NULL \code{dateNm} will be used.}

\item{kCategories}{If a categorical variable has more than kCategories, only 
a global histogram will be plotted, rate plots for all categories will also be
plotted.}

\item{normBy}{"time" or "var"-- the normalization factor for rate plots. If "time" then 
line plots will be normalized by month, that is, all caltegories will sum to one within
 a particular month.  This allows you to compare the proportion of total coming from 
different categories at 
different points in time. For example, at month 1 10% volume came from category A and at 
month 6 75%. If total volume was constant or decreasing, this represents an increase 
in market share of category A. If "var", then each category is normalized independently. 
In the same example, if category A shows a decreasing trend using the "var" normalization,
this would imply that total volume is decreasing and category A market share is increasing. 
This example occurs for the "retired" job category in the bankData example data set.}
}
\value{
Histogram and rate charts (if less than kCategories) for categorical data.
}
\description{
Frequency and proportion plots for discrete variables
}
\section{License}{

Copyright 2016 Capital One Services, LLC Licensed under the Apache License,
Version 2.0 (the "License"); you may not use this file except in compliance
with the License. You may obtain a copy of the  License at
http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law
or agreed to in writing, software distributed under the License is 
distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY 
KIND, either express or implied. See the License for the specific language 
governing permissions and limitations under the License.
}

\examples{
require(data.table)
data(bankData)
setDT(bankData)
require(ggplot2)

PrepData(bankData, dateNm = "date", dateGp = "months", dateGpBp = "quarters", 
         weightNm = NULL)
# Single histogram is plotted for job type since there are 12 categories
plot(PlotDiscreteVar(myVar = "job", dataFl = bankData, weightNm =  NULL, 
                     dateNm = "date", dateGp = "months"))
                     
plot(PlotDiscreteVar(myVar = "job", dataFl = bankData, 
                     weightNm = NULL, dateNm = "date", dateGp = "months",
                     kCategories = 12))


# binary data is treated as categorical, and only the less frequent category 
# is plotted over time
plot(PlotDiscreteVar(myVar = "default", dataFl = bankData, weightNm = NULL, 
                     dateNm = "date", dateGp = "months"))
}
