% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otvPlotsFunctions.R
\name{PlotHistogram}
\alias{PlotHistogram}
\title{Plot Histogram of Discrete Variable}
\usage{
PlotHistogram(glbTotals, myVar, newLevels = NULL)
}
\arguments{
\item{glbTotals}{A data.table with global counts of each category of \code{myVar}, 
produced by \code{\link{PlotDiscreteVar}}}

\item{myVar}{Name of the variable to be plotted}

\item{newLevels}{An optional vector naming the categories of \code{myVar} in 
order of frequency}
}
\value{
A ggproto object with a histogram of \code{myVar} ordered by category frequency
}
\description{
Plot Histogram of Discrete Variable
}
\section{License}{

Copyright 2016 Capital One Services, LLC Licensed under the Apache License,
Version 2.0 (the "License"); you may not use this file except in compliance
with the License. You may obtain a copy of the  License at
http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law
or agreed to in writing, software distributed under the License is 
distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY 
KIND, either express or implied. See the License for the specific language 
governing permissions and limitations under the License.
}
\examples{
data(bankData)
setDT(bankData)
PrepData(bankData, dateNm = "date", dateGp = "months", dateGpBp = "quarters", 
         weightNm = NULL)
glbTotals <- bankData[, .(count = .N), by = "job"]
PlotHistogram(glbTotals, "job")

#--- Some work can be pre-computed 
glbTotals <- bankData[, .(count = .N), by = "job"]
newLevels <- glbTotals[, job][order(glbTotals[, count])]
newLevels <- rev(unlist(newLevels))
PlotHistogram(glbTotals, "job", newLevels)
glbTotals[, ("job") := factor(job, levels = newLevels)]
PlotHistogram(glbTotals, "job", newLevels)
}

