% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otvPlotsFunctions.R
\name{PlotMean}
\alias{PlotMean}
\title{Plot mean with {Mean +- 1SD} control limits}
\usage{
PlotMean(meltdx, myVar, dateGp)
}
\arguments{
\item{meltdx}{A data.table with Mean and 1SD control limits in long format, 
produced by \code{\link{PlotVar}}}

\item{myVar}{Name of the variable to be plotted}

\item{dateGp}{Name of the variable the time series plots should be grouped
by. Options are NULL, "weeks", "months", "quarters", "years". See
data.table::IDate. If NULL \code{dateNm} will be used.}
}
\value{
A ggplot object with \code{dateGp} on the x axis, \code{value} on the
y axis, and variables \code{Mean}, \code{cl1} and \code{cl2} plotted on the
same graph, Mean and control limits differentiated by linetype
}
\description{
Plot mean with {Mean +- 1SD} control limits
}
\section{License}{
 
Copyright 2016 Capital One Services, LLC Licensed under the Apache License,
Version 2.0 (the "License"); you may not use this file except in compliance
with the License. You may obtain a copy of the  License at
http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law
or agreed to in writing, software distributed under the License is 
distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY 
KIND, either express or implied. See the License for the specific language 
governing permissions and limitations under the License.
}
\examples{
data(bankData)
setDT(bankData)
bankData[, months := round(date, "months")]
bankDT = bankData[, .( Mean = mean(balance)), by = "months"]
cl = bankData[, c(mean(balance), sd(balance))]
cl = cl \%*\% matrix(c(1, 1, 1, -1), byrow = TRUE, nrow = 2) # mean +- 1 SD
bankDT[, c("cl1", "cl2") := list(cl[1], cl[2])  ]
bankMT = melt(bankDT, id.vars = "months", 
         measure.vars = c("Mean", "cl1", "cl2"))
# bankMT # Long format for plotting
PlotMean(bankMT, "balance", "months")
}

