% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerical.R
\name{PlotQuantiles}
\alias{PlotQuantiles}
\title{Plot 01, 50, and 99 percentile for a numerical variable}
\usage{
PlotQuantiles(meltdx, myVar, dateGp)
}
\arguments{
\item{meltdx}{A data.table with p1, p50, and p99 in long format, produced by
\code{\link{SummaryStats}}}

\item{myVar}{The name of the variable to be plotted}

\item{dateGp}{Name of the variable that the time series plots should be 
grouped by. Options are \code{NULL}, \code{"weeks"}, \code{"months"}, 
\code{"quarters"}, \code{"years"}. See \code{\link[data.table]{IDate}} for
details. If \code{NULL}, then \code{dateNm} will be used as \code{dateGp}.}
}
\value{
A \code{ggplot2} object with \code{dateGp} on the x axis, 
  \code{value} on the y axis, and variables \code{p01}, \code{p50}, and 
  \code{p99} plotted on the same graph, with grouped and global percentiles 
  differentiated by linetype.
}
\description{
Plot 01, 50, and 99 percentile for a numerical variable
}
\section{License}{

Copyright 2016 Capital One Services, LLC Licensed under the Apache License,
Version 2.0 (the "License"); you may not use this file except in compliance
with the License. You may obtain a copy of the  License at
http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law
or agreed to in writing, software distributed under the License is 
distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY 
KIND, either express or implied. See the License for the specific language 
governing permissions and limitations under the License.
}

