% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{PlotRatesOverTime}
\alias{PlotRatesOverTime}
\title{Creates trace plots of categories' propotions over time for a discrete (or
binary) variable}
\usage{
PlotRatesOverTime(dataFl, dateGp, myVar, normBy = "time", weightNm = NULL,
  newLevels = NULL, kCategories = 9)
}
\arguments{
\item{dataFl}{Either the name of an object that can be converted using
\code{\link[data.table]{as.data.table}} (e.g., a data frame), or a 
character string containing the name of dataset that can be loaded using 
\code{\link[data.table]{fread}} (e.g., a csv file), or a file path  of Rdata
file. If dataset is not in your working directory then \code{dataFl}
must include (relative or absolute) path to file.}

\item{dateGp}{Name of the variable the time series plots should be grouped
by. Options are \code{NULL}, \code{"weeks"}, \code{"months"}, 
\code{"quarters"}, \code{"years"}. See \code{\link[data.table]{IDate}} for
details. If \code{NULL}, then \code{dateNm} will be used as \code{dateGp}.}

\item{myVar}{Name of the variable to be plotted}

\item{normBy}{The normalization factor for rate plots, can be \code{"time"}
or \code{"var"}. If \code{"time"}, then for each time period of 
\code{dateGp}, counts are normalized by the total counts over all 
categories in that time period. This illustrates changes of categories' 
proportions over time. If \code{"var"}, then for each category, its counts 
are normalized by the total counts over time from only this category. This
illustrates changes of categories' volumns over time.}

\item{weightNm}{Name of variable containing row weights, or \code{NULL} for 
no weights (all rows recieve weight 1).}

\item{newLevels}{categories of \code{myVar} in order of global frequency}

\item{kCategories}{If a categorical variable has more than \code{kCategories},
trace plots of only the \code{kCategories} most prevalent categories will
be plotted.}
}
\value{
A \code{ggplot} object, trace plots of categories' propotions 
  \code{myVar} over time.
}
\description{
Creates trace plots of categories' propotions over time for a discrete (or
binary) variable
}
\section{License}{

Copyright 2016 Capital One Services, LLC Licensed under the Apache License,
Version 2.0 (the "License"); you may not use this file except in compliance
with the License. You may obtain a copy of the  License at
http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law
or agreed to in writing, software distributed under the License is 
distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY 
KIND, either express or implied. See the License for the specific language 
governing permissions and limitations under the License.
}

\examples{
data(bankData)
setDT(bankData)
bankData[, weight := rpois(.N, 5)]
bankData[, weight := weight/sum(weight)]
bankData = PrepData(bankData, dateNm = "date", dateGp = "months", 
                    dateGpBp = "quarters", weightNm = "weight")
PlotRatesOverTime(dataFl = bankData, dateGp = "months", weightNm = "weight",
                  myVar = "job", newLevels = NULL, normBy = "time")
PlotRatesOverTime(dataFl = bankData, dateGp = "months",  weightNm = "weight",
                  myVar = "job", newLevels = NULL, normBy = "var")

}
