% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otvPlotsFunctions.R
\name{PrepData}
\alias{PrepData}
\title{Clean an input dataset for plotting}
\usage{
PrepData(dataFl, dateNm, selectCols = NULL, dropCols = NULL,
  dateFt = "\%d\%h\%Y", dateGp = NULL, dateGpBp = NULL, weightNm = NULL,
  varNms = NULL, dropConstants = TRUE, ...)
}
\arguments{
\item{dataFl}{Either the name of an object that can be converted using
as.data.table (e.g. a ' data frame) or a character string containing the name
of dataset that can be loaded using fread (e.g. a csv file). ' If dataset is
not in your working directory then \code{dataFl} must include (relative or 
absolute) path to file}

\item{dateNm}{Name of column containing date variable}

\item{selectCols}{Either NULL, or a vector contaning names or indices of
variables to read into memory -- must include \code{dateNm},
\code{weightNm} (if not null) and all variables to be plotted. If both
selectCols and dropCols are null, then all variables will be read in. Only
used when dataFl is a string constant, because selectCols is passed as a
parameter to fread. When dataFl is already a dataset in memory, use varNms
to indicate the column names to be plotted.}

\item{dropCols}{Either NULL or a vector of variables not to read into memory.
This parameter is passed directly to fread (see selectCols)}

\item{dateFt}{strptime format of date variable. Default is SAS format. See
?strptime}

\item{dateGp}{Name of the variable the time series plots should be grouped
by. Options are NULL, "weeks", "months", "quarters", "years". See
data.table::IDate. If NULL \code{dateNm} will be used.}

\item{dateGpBp}{Name of variable the boxplots should be grouped by. Same
options as \code{dateGp}. If NULL \code{dateGp} will be used.}

\item{weightNm}{Name of variable containing row weights or NULL for no
weights (all rows recieve weight 1)}

\item{varNms}{Either NULL or a vector of column names to be plotted. If null,
will default to all columns which are not \code{dateNm} or \code{weightNm}.
Can also be a vector of indices of the column names, after dropCols or
selectCols have been applied, if applicable, and not including dateGp,
dateGpBp (which will be added to the data.table)}

\item{dropConstants}{logical Indicates whether or not constant (all
duplicated or NA) variables should be dropped from dataFl prior to plotting}

\item{...}{Additional parameters to be passed to fread}
}
\value{
A data table formated for use by \code{PlotVar} function
}
\description{
Cleans an input dataset for use by otvPlots. \code{dataFl} must contain, at a
minimum, a date column \code{dateNm} and a variable to be plotted.
}
\section{License}{

Copyright 2016 Capital One Services, LLC Licensed under the Apache License,
Version 2.0 (the "License"); you may not use this file except in compliance
with the License. You may obtain a copy of the  License at
http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law
or agreed to in writing, software distributed under the License is 
distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY 
KIND, either express or implied. See the License for the specific language 
governing permissions and limitations under the License.
}
\examples{
data(bankData)
setDT(bankData)
PrepData(bankData, dateNm = "date", dateGp = "months", dateGpB = "quarters")
# view result
bankData   
# columns have been assigned a plotting class (cntns/dscrt)
str(bankData) 

}
\seealso{
\code{\link{fread}}
}

