% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otvPlotsFunctions.R
\name{SummaryStats}
\alias{SummaryStats}
\title{Create data.table of summary statistics for continuous plotting functions}
\usage{
SummaryStats(myVar, dataFl, dateGp, weightNm = NULL)
}
\arguments{
\item{myVar}{Name of the variable to be plotted}

\item{dataFl}{Either the name of an object that can be converted using
as.data.table (e.g. a ' data frame) or a character string containing the name
of dataset that can be loaded using fread (e.g. a csv file). ' If dataset is
not in your working directory then \code{dataFl} must include (relative or 
absolute) path to file}

\item{dateGp}{Name of the variable the time series plots should be grouped
by. Options are NULL, "weeks", "months", "quarters", "years". See
data.table::IDate. If NULL \code{dateNm} will be used.}

\item{weightNm}{Name of variable containing row weights or NULL for no
weights (all rows recieve weight 1)}
}
\value{
A data.table formatted for use by the continuous plot funtions 
\code{PlotMean}, \code{PlotQuantiles} and \code{PlotRates}.
}
\description{
Create data.table of summary statistics for continuous plotting functions
}
\section{License}{

Copyright 2016 Capital One Services, LLC Licensed under the Apache License,
Version 2.0 (the "License"); you may not use this file except in compliance
with the License. You may obtain a copy of the  License at
http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law
or agreed to in writing, software distributed under the License is 
distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY 
KIND, either express or implied. See the License for the specific language 
governing permissions and limitations under the License.
}
\examples{
data(bankData)
setDT(bankData)
PrepData(bankData, dateNm = "date", dateGp = "months", dateGpBp = "years")
mdx = SummaryStats(myVar = "age", dataFl = bankData, dateGp = "months")
plot(PlotQuantiles(mdx[variable \%in\% c("p99", "p50", "p1", "p99_g", 
                                            "p50_g", "p1_g")], "age", "months"))
plot(PlotMean(mdx[variable \%in\% c("Mean", "cl1", "cl2")], "age", "months"))
plot(PlotRates(mdx, "age", "months"))
}

