% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bankData}
\alias{bankData}
\title{The data is related with direct marketing campaigns of a Portuguese banking institution.}
\format{A data frame with 45,211 rows and 19 variables:
\describe{
  \item{age}{numeric}
  \item{job}{type of job (categorical: 'admin.','blue-collar','entrepreneur','housemaid','management','retired','self-employed','services','student','technician','unemployed','unknown')}
  \item{marital}{marital status (categorical: 'divorced','married','single','unknown'; note: 'divorced' means divorced or widowed)}
  \item{education}{categorical: 'basic.4y','basic.6y','basic.9y','high.school','illiterate','professional.course','university.degree','unknown')}
  \item{default}{has credit in default? (categorical: 'no','yes','unknown')}
  \item{balance}{account balance}
  \item{housing}{has housing loan? (categorical: 'no','yes','unknown')}
  \item{loan}{has personal loan? (categorical: 'no','yes','unknown')}
  \item{contact}{contact communication type (categorical: 'cellular','telephone')}
  \item{duration}{last contact duration, in seconds (numeric)}
  \item{campaign}{number of contacts performed during this campaign and for this client (numeric, includes last contact)}
  \item{pdays}{number of days that passed by after the client was last contacted from a previous campaign (numeric; 999 means client was not previously contacted)}
  \item{previous}{number of contacts performed before this campaign and for this client (numeric)}
  \item{poutcome}{outcome of the previous marketing campaign (categorical: 'failure','nonexistent','success')}
  \item{y}{has the client subscribed a term deposit? (binary: 'yes','no')}
  \item{date}{last contact date}
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Bank+Marketing}

\cite{Lichman, M. (2013). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science.}

\cite{[Moro et al., 2014] S. Moro, P. Cortez and P. Rita. A Data-Driven Approach to Predict the Success of Bank Telemarketing. Decision Support Systems, Elsevier, 62:22-31, June 2014}
}
\usage{
bankData
}
\description{
The marketing campaigns were based on phone calls.
Often, more than one contact to the same client was required, in order to access if the product (bank term deposit)
would be ('yes') or not ('no') subscribed.
Examples are ordered by date (from May 2008 to November 2010), similar to data analyzed in [Moro et al., 2014]
}
\keyword{datasets}

