% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bankLabels}
\alias{bankLabels}
\title{Labels for bankData}
\format{A data frame with 16 rows and 3 variables:
\describe{
  \item{V1}{Name of each variable in \code{\link{bankData}}.}
  \item{V2}{Label of each variable in \code{\link{bankData}}.}
  \item{V3}{A numeric variable, corresponding to the row number.}
}}
\usage{
bankLabels
}
\description{
A dataset containing the attribute labels also found in \code{\link{bankData}}.
This data set is used to illustrate the \code{\link{PrepLabels}} function and
other label functionality in the \code{\link{otvPlots}} package in R.
}
\keyword{datasets}
