% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{PlotBarplot}
\alias{PlotBarplot}
\title{Creates a bar plot for a discrete (or binary) variable}
\usage{
PlotBarplot(dataFl, myVar, weightNm = NULL)
}
\arguments{
\item{dataFl}{Either the name of an object that can be converted using
\code{\link[data.table]{as.data.table}} (e.g., a data frame), or a 
character string containing the name of dataset that can be loaded using 
\code{\link[data.table]{fread}} (e.g., a csv file), or a file path  of Rdata
file. If dataset is not in your working directory then \code{dataFl}
must include (relative or absolute) path to file.}

\item{myVar}{Name of the variable to be plotted}

\item{weightNm}{Name of variable containing row weights, or \code{NULL} for 
no weights (all rows recieve weight 1).}
}
\value{
A \code{ggplot} object with a histogram of \code{myVar} ordered by 
  category frequency
}
\description{
Creates a bar plot for a discrete (or binary) variable
}
\section{License}{

Copyright 2016 Capital One Services, LLC Licensed under the Apache License,
Version 2.0 (the "License"); you may not use this file except in compliance
with the License. You may obtain a copy of the  License at
http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law
or agreed to in writing, software distributed under the License is 
distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY 
KIND, either express or implied. See the License for the specific language 
governing permissions and limitations under the License.
}

\examples{
data(bankData)
bankData = PrepData(bankData, dateNm = "date", dateGp = "months", 
                    dateGpBp = "quarters", weightNm = NULL)
PlotBarplot(bankData, "job")

# NA will be included as a category if any NA are present
bankData[sample.int(.N)[1:1000], education := NA]
PlotBarplot(bankData, "education")
}
