% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otvPlotsFunctions.R
\name{PlotCatRate}
\alias{PlotCatRate}
\title{Plot Category Frequency Rates Over Time}
\usage{
PlotCatRate(rateData, myVar, dateGp, hex = NULL)
}
\arguments{
\item{rateData}{A data.table with proportion of \code{myVar} following into the most
frequent (globally) kCategories of \code{myVar} per time period (\code{dateGp}), 
produced by \code{\link{PlotDiscreteVar}}, optionally weighted by \code{weightNm}}

\item{myVar}{Name of the variable to be plotted}

\item{dateGp}{Name of the variable the time series plots should be grouped
by. Options are NULL, "weeks", "months", "quarters", "years". See
data.table::IDate. If NULL \code{dateNm} will be used.}
}
\value{
A ggproto object with plotting the proportion of \code{myVar} per category over time, 
for the most frequent (globally) \code{kCategories} of \code{myVar}
}
\description{
Plot Category Frequency Rates Over Time
}
\section{License}{

Copyright 2016 Capital One Services, LLC Licensed under the Apache License,
Version 2.0 (the "License"); you may not use this file except in compliance
with the License. You may obtain a copy of the  License at
http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law
or agreed to in writing, software distributed under the License is 
distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY 
KIND, either express or implied. See the License for the specific language 
governing permissions and limitations under the License.
}
\examples{
data(bankData)
setDT(bankData)
PrepData(bankData, dateNm = "date", dateGp = "months", dateGpBp = "quarters", 
         weightNm = NULL)
jobTable = rev(sort(bankData[, table(job)]))
jobTable1 = order(bankData[, table(job)])
nLevels = length(jobTable)
kCategories = 3
topLevels <- names(jobTable)[1:kCategories]
rateData <- bankData[, .N, by = c("job", "months")][job \%in\% topLevels]
rate1    <- bankData[, .N, by = "months"]
rateData <- merge(rateData, rate1, by = "months")
rateData[, rate := N.x / N.y]
rateData[, ("job") := factor(job, levels = topLevels)]
PlotCatRate(rateData, "job", "months", hex = NULL)
# you can specify the colours of the plot

PlotCatRate(rateData, "job", "months", hex = c("#00BA38", "#00BFC4", "#619CFF"))
}

