% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otvPlotsFunctions.R
\name{PlotDiscreteVar}
\alias{PlotDiscreteVar}
\title{Frequency and proportion plots for discrete variables}
\usage{
PlotDiscreteVar(myVar, dataFl, weightNm, dateNm, dateGp, kCategories = 3,
  refactorInd = FALSE)
}
\arguments{
\item{myVar}{Name of the variable to be plotted}

\item{dataFl}{A data.table containing at least the following column names:
\code{myVar}, \code{weightNm}, \code{dateGp}, \code{dateGpBp}. Output of
PrepData}

\item{weightNm}{Name of variable containing row weights or NULL for no
weights (all rows recieve weight 1)}

\item{dateNm}{Name of column containing date variable}

\item{dateGp}{Name of the variable the time series plots should be grouped
by. Options are NULL, "weeks", "months", "quarters", "years". See
data.table::IDate. If NULL \code{dateNm} will be used.}

\item{kCategories}{Number of categories to plot in over time
rate chart for categorical data -- the K most numerous categories in the 
dataset will be plotted. Default is 3.}

\item{refactorInd}{Logical. Indicates whether discrete 
\code{myVar} should be converted to factor as plotting time. Doing so will 
ensure over time histogram plots are in the same order as other discrete 
plots, but will substantially increase processing time for large data sets}
}
\value{
Histograms and rate charts for categorical data. For factors with <= 
9 levels both a bar chart and barchart over time are returned. A single 
global histogram is returned for variables with more than 9 categories.
}
\description{
Frequency and proportion plots for discrete variables
}
\section{License}{

Copyright 2016 Capital One Services, LLC Licensed under the Apache License,
Version 2.0 (the "License"); you may not use this file except in compliance
with the License. You may obtain a copy of the  License at
http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law
or agreed to in writing, software distributed under the License is 
distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY 
KIND, either express or implied. See the License for the specific language 
governing permissions and limitations under the License.
}
\examples{
require(data.table)
data(bankData)
setDT(bankData)
require(ggplot2)

PrepData(bankData, dateNm = "date", dateGp = "months", dateGpBp = "quarters", 
         weightNm = NULL)
# Single histogram is plotted for job type since there are 12 categories
plot(PlotDiscreteVar(myVar = "job", dataFl = bankData, weightNm =  NULL, 
                     dateNm = "date", dateGp = "months"))
# Limit to top 9 job categories and plot over time
topJobs = rev(names(sort(table(bankData[, job]))))[1:9]
plot(PlotDiscreteVar(myVar = "job", dataFl = bankData[job \%in\% topJobs], 
                     weightNm = NULL, dateNm = "date", dateGp = "months"))

# By default only the top 3 categories are plotted in the rate graph, but we
# can have all categegories plotted. 
# By setting refactorInd = TRUE, the histogram is put in order of
# global volume and the colors will match the over time rate plot. However, 
# this option can dramatically increase processing time in a large dataset. 
plot(PlotDiscreteVar(myVar = "job", dataFl = bankData[job \%in\% topJobs], 
                     weightNm = NULL, dateNm = "date", dateGp = "months", 
                     kCategories = 9, refactorInd = TRUE)) 

#  binary data only gets two plots
plot(PlotDiscreteVar(myVar = "housing", dataFl = bankData, weightNm = NULL, 
                     dateNm = "date", dateGp = "months"))
}

