% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otvPlotsFunctions.R
\name{PlotHistogram}
\alias{PlotHistogram}
\title{Plot Histogram of Discrete Variable}
\usage{
PlotHistogram(dataFl, myVar, weightNm = NULL)
}
\arguments{
\item{dataFl}{Either the name of an object that can be converted using
as.data.table (e.g. a ' data frame), or a character string containing the name
of dataset that can be loaded using fread (e.g. a csv file), or a file path of Rdata file. ' If dataset is
not in your working directory then \code{dataFl} must include (relative or 
absolute) path to file}

\item{myVar}{Name of the variable to be plotted}

\item{weightNm}{Name of variable containing row weights or NULL for no
weights (all rows recieve weight 1)}
}
\value{
A ggproto object with a histogram of \code{myVar} ordered by category frequency
}
\description{
Plot Histogram of Discrete Variable
}
\section{License}{

Copyright 2016 Capital One Services, LLC Licensed under the Apache License,
Version 2.0 (the "License"); you may not use this file except in compliance
with the License. You may obtain a copy of the  License at
http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law
or agreed to in writing, software distributed under the License is 
distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY 
KIND, either express or implied. See the License for the specific language 
governing permissions and limitations under the License.
}

\examples{
data(bankData)
setDT(bankData)
PrepData(bankData, dateNm = "date", dateGp = "months", dateGpBp = "quarters", 
         weightNm = NULL)
PlotHistogram(bankData, "job")

# NA will be included as a category if any NA are present
bankData[sample.int(.N)[1:1000], education := NA]
PlotHistogram(bankData, "education")
}
