% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otvPlotsFunctions.R
\name{PlotHistOverTime}
\alias{PlotHistOverTime}
\title{Plot Histogram of Discrete Variable Over Time}
\usage{
PlotHistOverTime(dataFl, dateGp, myVar, normBy = "time", weightNm = NULL,
  newLevels = NULL)
}
\arguments{
\item{dataFl}{Either the name of an object that can be converted using
as.data.table (e.g. a ' data frame), or a character string containing the name
of dataset that can be loaded using fread (e.g. a csv file), or a file path of Rdata file. ' If dataset is
not in your working directory then \code{dataFl} must include (relative or 
absolute) path to file}

\item{dateGp}{Name of the variable the time series plots should be grouped
by. Options are NULL, "weeks", "months", "quarters", "years". See
data.table::IDate. If NULL \code{dateNm} will be used.}

\item{myVar}{Name of the variable to be plotted}

\item{normBy}{"time" or "var"-- the normalization factor for rate plots. If "time" then 
line plots will be normalized by month, that is, all caltegories will sum to one within
 a particular month.  This allows you to compare the proportion of total coming from 
different categories at 
different points in time. For example, at month 1 10% volume came from category A and at 
month 6 75%. If total volume was constant or decreasing, this represents an increase 
in market share of category A. If "var", then each category is normalized independently. 
In the same example, if category A shows a decreasing trend using the "var" normalization,
this would imply that total volume is decreasing and category A market share is increasing. 
This example occurs for the "retired" job category in the bankData example data set.}

\item{weightNm}{Name of variable containing row weights or NULL for no
weights (all rows recieve weight 1)}

\item{newLevels}{categories of myVar in order of global frequency}
}
\value{
A ggproto object with a histogram of \code{myVar} over time
}
\description{
Plot Histogram of Discrete Variable Over Time
}
\section{License}{

Copyright 2016 Capital One Services, LLC Licensed under the Apache License,
Version 2.0 (the "License"); you may not use this file except in compliance
with the License. You may obtain a copy of the  License at
http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law
or agreed to in writing, software distributed under the License is 
distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY 
KIND, either express or implied. See the License for the specific language 
governing permissions and limitations under the License.
}

\examples{
data(bankData)
setDT(bankData)
bankData[, weight := rpois(.N, 5)]
bankData[, weight := weight/sum(weight)]
PrepData(bankData, dateNm = "date", dateGp = "months", dateGpBp = "quarters", 
         weightNm = "weight")
PlotHistOverTime(dataFl = bankData, dateGp = "months", 
weightNm = "weight", myVar = "job", newLevels = NULL, normBy = "time")
PlotHistOverTime(dataFl = bankData, dateGp = "months", 
weightNm = "weight", myVar = "job", newLevels = NULL, normBy = "var")

}
