% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otvPlotsFunctions.R
\name{PlotQuantiles}
\alias{PlotQuantiles}
\title{Plots  01, 50, and 99 percentile together}
\usage{
PlotQuantiles(meltdx, myVar, dateGp)
}
\arguments{
\item{meltdx}{A data.table with p1, p50, and p99 in long format, produced by 
\code{\link{PlotVar}}}

\item{myVar}{Name of the variable to be plotted}

\item{dateGp}{Name of the variable the time series plots should be grouped by. Options are NULL, 
"weeks", "months", "quarters", "years". See data.table::IDate. If NULL \code{dateNm} will be
used.}
}
\value{
A ggplot object with \code{dateGp} on the x axis, \code{value} on the y axis,
and variables \code{p01}, \code{p50} and \code{p99} plotted on the same graph, 
with grouped and global percentiles differentiated by linetype
}
\description{
Plots  01, 50, and 99 percentile together
}
\section{License}{

Copyright 2016 Capital One Services, LLC Licensed under the Apache License, Version 2.0 
(the "License");  you may not use this file except in compliance with the License. 
You may obtain a copy of the  License at http://www.apache.org/licenses/LICENSE-2.0
Unless required by applicable law or agreed to in writing, software distributed under the 
License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
either express or implied. See the License for the specific language governing permissions 
and limitations under the License.
}
\examples{
data(bankData)
setDT(bankData)
bankData[, months := round(date, "months")]
bankDT = bankData[, {
  tmp1=quantile(balance, p=c(.01, .5, .99) );
 list("p1"  = tmp1[1] ,
      "p50" = tmp1[2] ,
      "p99" = tmp1[3]
 ) } , by   = "months"]

bankMT = melt(bankDT, id.vars="months", measure.vars=c("p99", "p50","p1"))
globalPct = bankData[ , quantile(balance, p=c(.01, .5, .99) ) ]
globalDT = data.table("months" = rep(bankMT[variable=="p99", "months", with=FALSE][[1]],3))
globalDT[, c("variable", "value"):= list(rep(c("p1_g", "p50_g", "p99_g"), each=.N/3), 
  rep(globalPct, each=.N/3))]
bankMT= rbindlist(list( bankMT, globalDT))
#---- Long format for plotting
bankMT
PlotQuantiles(bankMT, "balance", "months")
}

